[python] Add and implement a basic mono delay effect
This delay was based on the delay explained in this article:
https://wiki.analog.com/resources/tools-software/sharc-audio-module/baremetal/delay-effect-tutorial

It uses a simple delay line buffer that cycles based on the delay time.

First, the new output can be determined by mixing the contents of the
next chunk of the delay buffer with the input audio.

Second, the delay buffer is updated with the next chunk of input audio.
Inserted audio wraps around the delay buffer with the same period as the
delay time (delaying the effect the audio signal has on the output by the
time it takes to wrap around the delay buffer again).

Finally, the feedback multiple scales the inserted audio chunk in the
delay buffer.

Test command:
```
python main.py --sample-rate 48000 delay ../wavs/click.wav delay.wav --time 0.05
```